requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
langchain==0.0.155
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
pypdf2==3.0.1
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.2
requests==2.29.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
uvicorn==0.22.0
yarl==1.9.2
